requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
tqdm
matplotlib
Pillow
